bcachefs: Fix for freespace version upgrade path
It's currently possible to end up in a half-upgraded state where we
haven't set the superblock to the new version, but we have run the
freespace initialization path. Previously, this meant when running fsck
on such a filesystem we wouldn't check the freespace btrees - which is a
problem, if they have been initialized and there's something fsck needs
to check and fix.

Fix this by making bch2_check_alloc_info() check if freespace has been
initialized on each device, not by making it run conditionally on the
superblock version.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>